Fix the Uin parse error and log file error

+++ wechat/wechat_params.go
@@ -85,7 +85,7 @@ type BaseRequest struct {
 	Message    string   `xml:"message" json:"-"`
 	Skey       string   `xml:"skey" json:"Skey"`
 	Wxsid      string   `xml:"wxsid" json:"Sid"`
-	Wxuin      int      `xml:"wxuin" json:"Uin"`
+	Wxuin      int64    `xml:"wxuin" json:"Uin"`
 	PassTicket string   `xml:"pass_ticket" json:"-"`
 	DeviceID   string   `xml:"-" json:"DeviceID"`
 }
@@ -156,7 +156,7 @@ func (this *Member) IsSpecail() bool {
 
 type User struct {
 	UserName          string `json:"UserName"`
-	Uin               int    `json:"Uin"`
+	Uin               int64  `json:"Uin"`
 	NickName          string `json:"NickName"`
 	HeadImgUrl        string `json:"HeadImgUrl" xml:""`
 	RemarkName        string `json:"RemarkName" xml:""`
@@ -177,7 +177,7 @@ type User struct {
 }
 
 type Member struct {
-	Uin              int
+	Uin              int64
 	UserName         string
 	NickName         string
 	HeadImgUrl       string
@@ -189,7 +189,7 @@ type Member struct {
 	Sex              int
 	Signature        string
 	VerifyFlag       int
-	OwnerUin         int
+	OwnerUin         int64
 	PYInitial        string
 	PYQuanPin        string
 	RemarkPYInitial  string
